About me done, organisational changes.
I finished the about me page for now, added another image and finished the text. It will definitely be changed again but I'm gonna try to focus on finishing the rest first, also edited the changelist on the main page.

Also changed the organisation of my files and images to make them less messy.